+++ packages/gestalt/src/touchable.jsdom.test.js
@@ -12,6 +12,25 @@ test('Touchable handles onTouch callback', () => {
   expect(mockOnTouch).toHaveBeenCalled();
 });
 
+test('Touchable handles onBlur callback', () => {
+  const mockOnBlur = jest.fn();
+  const { getByText } = render(
+    <Touchable onBlur={mockOnBlur}>Touchable</Touchable>
+  );
+  fireEvent.focus(getByText('Touchable'));
+  fireEvent.blur(getByText('Touchable'));
+  expect(mockOnBlur).toHaveBeenCalled();
+});
+
+test('Touchable handles onFocus callback', () => {
+  const mockOnFocus = jest.fn();
+  const { getByText } = render(
+    <Touchable onFocus={mockOnFocus}>Touchable</Touchable>
+  );
+  fireEvent.focus(getByText('Touchable'));
+  expect(mockOnFocus).toHaveBeenCalled();
+});
+
 test('Touchable handles onMouseEnter callback', () => {
   const mockOnMouseEnter = jest.fn();
   const { getByText } = render(